refactor(DatePicker): Abstract selection logic into hook
