+++ packages/react-component-library/src/components/DatePicker/types.ts
@@ -1,0 +1,4 @@
+export interface StateObject {
+  from?: Date
+  to?: Date
+}
